+++ servant-server/test/Servant/Server/ErrorSpec.hs
@@ -6,32 +6,38 @@
 module Servant.Server.ErrorSpec (spec) where
 
 import           Data.Aeson                 (encode)
-import qualified Data.ByteString.Lazy.Char8 as BC
+import qualified Data.ByteString.Lazy.Char8 as BCL
+import qualified Data.ByteString.Char8      as BC
 import           Data.Proxy
 import           Network.HTTP.Types         (hAccept, hContentType, methodGet,
-                                             methodPost)
+                                             methodPost, methodPut)
 import           Test.Hspec
 import           Test.Hspec.Wai
 
 import           Servant
 
 
--- 1) Check whether one or more endpoints have the right path. Otherwise return 404.
--- 2) Check whether the one of those have the right method. Otherwise return
--- 405. If so, pick the first. We've now commited to calling at most one handler. *
--- 3) Check whether the Content-Type is known. Otherwise return 415.
--- 4) Check whether that one deserializes the body. Otherwise return 400. If there
--- was no Content-Type, try the first one of the API content-type list.
--- 5) Check whether the request is authorized. Otherwise return a 401.
--- 6) Check whether the request is forbidden. If so return 403.
--- 7) Check whether the request has a known Accept. Otherwise return 406.
--- 8) Check whether Accept-Language, Accept-Charset and Accept-Encoding exist and
--- match. We can follow the webmachine order here.
--- 9) Call the handler. Whatever it returns, we return.
+-- The semantics of routing and handling requests should be as follows:
+--
+--    1) Check whether one or more endpoints have the right path. Otherwise
+--    return 404.
+--    2) Check whether the one of those have the right method. Otherwise return
+--    405. If so, pick the first. We've now committed to calling at most one
+--    handler.
+--    3) Check whether the Content-Type is known. Otherwise return 415.
+--    4) Check whether that one deserializes the body. Otherwise return 400. If
+--    there was no Content-Type, try the first one of the API content-type list.
+--    5) Check whether the request is authorized. Otherwise return a 401.
+--    6) Check whether the request is forbidden. If so return 403.
+--    7) Check whether the request has a known Accept. Otherwise return 406.
+--    8) Check whether Accept-Language, Accept-Charset and Accept-Encoding
+--    exist and match. We can follow the webmachine order here.
+--    9) Call the handler. Whatever it returns, we return.
 
 spec :: Spec
 spec = describe "HTTP Errors" $ do
     errorOrderSpec
+    prioErrorsSpec
     errorRetrySpec
     errorChoiceSpec
 
@@ -82,6 +88,52 @@ errorOrderSpec = describe "HTTP error order"
   it "has 406 as its fifth highest priority error" $ do
     request goodMethod goodUrl [goodContentType, badAccept] goodBody
       `shouldRespondWith` 406
+
+type PrioErrorsApi = ReqBody '[JSON] Integer :> "foo" :> Get '[JSON] Integer
+
+prioErrorsApi :: Proxy PrioErrorsApi
+prioErrorsApi = Proxy
+
+-- Check whether matching continues even if a 'ReqBody' or similar construct
+-- is encountered early in a path. We don't want to see a complaint about the
+-- request body unless the path actually matches.
+prioErrorsSpec :: Spec
+prioErrorsSpec = describe "PrioErrors" $ do
+  let server = return
+  with (return $ serve prioErrorsApi server) $ do
+    let check (mdescr, method) path (cdescr, ctype, body) resp =
+          it fulldescr $
+            Test.Hspec.Wai.request method path [(hContentType, ctype)] body
+              `shouldRespondWith` resp
+          where
+            fulldescr = "returns " ++ show (matchStatus resp) ++ " on " ++ mdescr
+                     ++ " " ++ (BC.unpack path) ++ " (" ++ cdescr ++ ")"
+
+        get' = ("GET", methodGet)
+        put' = ("PUT", methodPut)
+
+        txt   = ("text"        , "text/plain;charset=utf8"      , "42"        )
+        ijson = ("invalid json", "application/json;charset=utf8", "invalid"   )
+        vjson = ("valid json"  , "application/json;charset=utf8", encode (5 :: Int))
+
+    check get' "/"    txt   404
+    check get' "/bar" txt   404
+    check get' "/foo" txt   415
+    check put' "/"    txt   404
+    check put' "/bar" txt   404
+    check put' "/foo" txt   405
+    check get' "/"    ijson 404
+    check get' "/bar" ijson 404
+    check get' "/foo" ijson 400
+    check put' "/"    ijson 404
+    check put' "/bar" ijson 404
+    check put' "/foo" ijson 405
+    check get' "/"    vjson 404
+    check get' "/bar" vjson 404
+    check get' "/foo" vjson 200
+    check put' "/"    vjson 404
+    check put' "/bar" vjson 404
+    check put' "/foo" vjson 405
 
 -- }}}
 ------------------------------------------------------------------------------
@@ -190,8 +242,10 @@ errorChoiceSpec = describe "Multiple handlers return errors"
 -- * Instances {{{
 
 instance MimeUnrender PlainText Int where
-    mimeUnrender _ = Right . read . BC.unpack
+    mimeUnrender _ = Right . read . BCL.unpack
 
 instance MimeRender PlainText Int where
-    mimeRender _ = BC.pack . show
+    mimeRender _ = BCL.pack . show
 -- }}}
+--
+